+++ source/reorder_audio_streams2/changelog.md
@@ -1,8 +1,10 @@
 
+**<span style="color:#56adda">0.0.11</span>**
+- add all pages from tmdb query
+
 **<span style="color:#56adda">0.0.10</span>**
 - rename plugin to reorder_audio_streams2
 - make reordering original languages an option
-- add all pages from tmdb query
 
 **<span style="color:#56adda">0.0.9</span>**
 - fix unique test function results so only used when more than 1 result found